Refactor, add auto analysis
