+++ ghidra-assistant/src/main/java/reva/RevaChatService.java
@@ -1,0 +1,6 @@
+package reva;
+
+public interface RevaChatService {
+    public String revaChat(String message);
+    public void blockUntilConnected();
+}
