+++ reverse-engineering-assistant/reverse_engineering_assistant/api_server_tools/re_tool_box/__init__.py
@@ -1,0 +1,6 @@
+"""
+To make sure your tool is loaded, you should import it here.
+`assistant_api_server.py` imports everything from `re_tools.py`,
+and `re_tools.py` imports everything from here.
+"""
+
